+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * Add command to generate LICENSE file (#155)
 * Add new DSLs `author` and `license` (#156)
 * Add license linter (#157)
+* Allow few commands' execution without any Eask-file (#159)
 
 ## 0.8.x
 > Released Mar 08, 2023